Incomplete - task 33: Add a (x,y) (time,speed) and a (x,y) (time,distance) graph to the stats screen","Type-Enhancement, Priority-Medium, Milestone-Release0.9 http://code.google.com/p/open-gpstracker/issues/detail?id=33
Save and restore state

+++ application/src/nl/sogeti/android/gpstracker/actions/utils/GraphCanvas.java
@@ -95,7 +95,7 @@ public class GraphCanvas extends View
       }
    }
    
-   public long getType()
+   public int getType()
    {
       return mGraphType;
    }
